BACKGROUND_PROJECT/PYTHON_PROJECT/requirements.txt
mysql-connector-python==8.0.25
flask==2.0.1
flask-login==0.5.0
Flask-SQLAlchemy




